refactored all index.js code into separate route js files for cleaner look
