+++ package.json
@@ -25,6 +25,7 @@
     "express-session": "^1.17.1",
     "jquery": "^3.5.1",
     "jsdom": "^16.4.0",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "nodemon": "^2.0.6"
   }
 }
