+++ routes/grade.js
@@ -1,0 +1,21 @@
+const express = require('express');
+const router = express.Router();
+
+
+/* global connection */
+router.get('/:gid', (req, res) => {
+    let stmt = "SELECT * FROM gradeLvlInfo WHERE gradeLvl = " + req.params.gid + ";";
+    console.log(stmt);
+    connection.query(stmt, function(err, result) {
+        if (err) throw err;
+
+        if (result.length) {
+            var gradeInfo = result[0];
+            console.log(gradeInfo);
+            res.render('grade', { gradeInfo: gradeInfo });
+        }
+    })
+});
+
+
+module.exports = router;
