+++ views/admin.ejs
@@ -27,7 +27,7 @@
             </div>
           </div>
         </div>
-        <form action='/adminadd' method='POST' class="addQform">
+        <form action='/admin/add' method='POST' class="addQform">
             <!--<div class="row">-->
             <!--    <div class="col-md-offset-1 col-md-2">-->
                     <h1>Add Question</h1>
@@ -57,7 +57,7 @@
             <!--</div>-->
         </form>
         <br><br>
-        <form action='/admindelete' method='POST' class="addQform">
+        <form action='/admin/delete' method='POST' class="addQform">
             <!--<div class="row">-->
             <!--    <div class="col-md-offset-1 col-md-2">-->
                     <h1>Delete Question Question</h1>
