+++ views/adminlogin.ejs
@@ -3,7 +3,7 @@
     <!-- Page Content -->
 <body id= "backgroundAdminLogin">
     <div class="container">
-    <form action='/adminlogin' method='POST'>
+    <form action='/admin/login' method='POST'>
         <div class="row">
             <div class="col-md-offset-4 col-md-4 text-center">
                   <div class="form-login"></br>
