+++ views/landing.ejs
@@ -19,7 +19,7 @@
 				        <button type="button" class="close" data-dismiss="modal">&times;</button>
 				      </div>
 				      <div class="modal-body" style="color: black;">
-				      	<form action="/gotoquiz" method="POST">
+				      	<form action="/quiz/startquiz" method="POST">
 				      		<label for="chooseDiff">Choose a difficulty:</label><br>
                             <input type="number" id="chooseDiff" name="chooseDiff" min="1" max="10" required><br>
                             <label for="chooseCat">Choose a category for your quiz:</label><br>
