+++ views/login.ejs
@@ -3,7 +3,7 @@
     <!-- Page Content -->
 <body id= "backgroundLogin">
     <div class="container">
-    <form action='/login' method='POST'>
+    <form action='/login/' method='POST'>
         <div class="row">
             <div class="col-md-offset-4 col-md-4 text-center">
                   <div class="form-login"></br>
