+++ views/templates/navbar.ejs
@@ -27,7 +27,7 @@
         <ul class="nav navbar-nav navbar-right">
           <li><a href='/register'><span class="glyphicon glyphicon-user"></span> Sign Up</a></li>
           <li><a href='/login'><span class="glyphicon glyphicon-log-in"></span> Login</a></li>
-          <li><a href='/adminlogin'><span class="glyphicon glyphicon-log-in"></span>Admin</a></li>
+          <li><a href='/admin/login'><span class="glyphicon glyphicon-log-in"></span>Admin</a></li>
         </ul>
       </div>
     </nav>
