+++ views/userprofile.ejs
@@ -42,7 +42,7 @@
                     
                     <div class="col-md-2">
                         <!--<input ty'pe="submit" class="profile-edit-btn" name="btnAddMore" value="Edit Profile"/>-->
-                        <a href='/editprofile' class='profile-edit-btn'>Edit Profile</a>
+                        <a href='/user/:uid/updateemail' class='profile-edit-btn'>Edit Profile</a>
                     </div>
